Moves the open web browser to the client
